requirements.txt
Django==1.6.2
djangorestframework==2.3.13
django-userena==1.3.0
